requirements.txt
flask
gunicorn
pandas
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
line-bot-sdk
openai
Pillow
requests
certifi
